Add support for (make check CHECK_BASE_IMAGES='fedora centos')
This is intended for use in Travis (or perhaps also locally).  The
default behavior remains the same, uses the fedora image.

(Note that the (yum install) does not fail when one of the packages
is missing.  This is not a dealbreaker, but it will make diagnosing
future bugs more difficult.)

Uses CentOS 7 because CentOS 8 is missing some necessary -devel packages.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>