Add support for (make check CHECK_BASE_IMAGES='fedora centos')
This is intended for use in Travis (or perhaps also locally).  The
default behavior remains the same, uses the fedora image.

Adds various workarounds for Go 1.4; we should drop them ASAP.

(Note that the (yum install) does not fail when one of the packages
is missing.  This is not a dealbreaker, but it will make diagnosing
future bugs more difficult.)